moving all user class into User-service
