+++ User-service/src/main/java/com/user/DTO/UserDto.java
@@ -1,4 +1,4 @@
-package com.task.DTO;
+package com.user.DTO;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
