+++ User-service/src/main/java/com/user/controllers/LoginController.java
@@ -1,4 +1,4 @@
-package com.task.controllers;
+package com.user.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
