+++ User-service/src/main/java/com/user/entities/Users.java
@@ -1,4 +1,4 @@
-package com.task.entities;
+package com.user.entities;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.persistence.*;
